Moved index.ts file and changed default language
