+++ src/DonateTo.WebApplication/ClientApp/src/app/app.component.ts
@@ -19,6 +19,6 @@ export class AppComponent {
     translate.setDefaultLang('en');
 
     // the lang to use, if the lang isn't available, it will use the current loader to get them
-    translate.use('es');
+    translate.use('en');
   }
 }
