+++ src/DonateTo.WebApplication/ClientApp/src/app/app.sandbox.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Sandbox } from './shared/sandbox/base.sandbox';
 import { Store } from '@ngrx/store';
-import * as store from './shared/store';
+import * as store from './shared/store/settings';
 import * as settingsActions from './shared/store/settings/settings.action';
 import { TranslateService } from 'ng2-translate';
 import { ConfigService } from './app-config.service';
